% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhaskara.R
\name{bhaskara}
\alias{bhaskara}
\title{Solução de Equações do Segundo Grau}
\usage{
bhaskara(a, b, c)
}
\arguments{
\item{a}{Coeficiente do termo quadrático.}

\item{b}{Coeficiente do termo linear.}

\item{c}{Coeficiente constante.}
}
\value{
Um vetor numérico de tamanho
0, 1 ou 2,
dependendo do número de soluções reais
encontradas.
}
\description{
Encontre a solução de equações de segundo
grau da forma
\deqn{ax^2 + bx + c = 0}
utilizando a fórmula de Bhaskara.
}
\details{
As soluções apresentadas são apenas as reais.
}
\examples{
bhaskara(1, 0, -1)
bhaskara(1, 0, 1)
bhaskara(4, 4, 1)

}
