% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados918}
\alias{dados918}
\title{Conjunto de dados do pacote}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
dados918
}
\description{
Dados gerados durante a aula
}
\keyword{datasets}
