% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello1.R, R/tchau.R
\name{hello}
\alias{hello}
\alias{tchau}
\title{Saudação}
\usage{
hello(nome, sobrenome)

tchau(nome, ate = "amanhã")
}
\arguments{
\item{nome}{Nome da pessoa.}

\item{sobrenome}{Sobrenome da pessoa.}

\item{ate}{Quando vai encontrar a pessoa.}

\item{aaaaa}{Erro?.}
}
\description{
Diz olá. Olhe também a fórmula
de \link[pacote1109]{bhaskara}.

Diz adeus.
}
